🔥 remove old iterator-based API
