+++ examples/visual_math/src/app/mod.rs
@@ -149,8 +149,7 @@ impl App {
                 }
             });
 
-        // let response = graph.show(ui);
-        let response = graph.show_using_visitor(ui);
+        let response = graph.show(ui);
 
         self.current_graph_pos = response.position;
     }
